For line number 193 & 194 | VNet peering Mismatch
!Note
Virtual network peering is available only for virtual networks in the same region.

In this document why it is mentioned as VNet peering is available only for VNets in the same region(same region implies to local VNet peering, isn't that correct?)
Global VNet peering is also supportable. 
So i believe here we need to change or make it clarify further more deep.